# contracts/openapi/openapi.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mutiAI Core API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/organizations/proposals": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Propose Organization",
        "operationId": "propose_organization_api_v1_organizations_proposals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganizationProposalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationVersionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "List Organizations",
        "operationId": "list_organizations_api_v1_organizations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizationSummaryResponse"
                  },
                  "type": "array",
                  "title": "Response List Organizations Api V1 Organizations Get"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get Organization",
        "operationId": "get_organization_api_v1_organizations__organization_id__get",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetailResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/versions": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "List Organization Versions",
        "operationId": "list_organization_versions_api_v1_organizations__organization_id__versions_get",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationVersionResponse"
                  },
                  "title": "Response List Organization Versions Api V1 Organizations  Organization Id  Versions Get"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/versions/{spec_version_id}/confirm": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Confirm Organization Version",
        "operationId": "confirm_organization_version_api_v1_organizations__organization_id__versions__spec_version_id__confirm_post",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization Id"
            }
          },
          {
            "name": "spec_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spec Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationVersionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/organizations/{organization_id}/versions/{spec_version_id}/publish": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Publish Organization Version",
        "operationId": "publish_organization_version_api_v1_organizations__organization_id__versions__spec_version_id__publish_post",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Organization Id"
            }
          },
          {
            "name": "spec_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Spec Version Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationVersionResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentRoleSpec": {
        "properties": {
          "role_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Role Key"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "responsibility": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Responsibility"
          },
          "is_lead": {
            "type": "boolean",
            "title": "Is Lead",
            "default": false
          },
          "reports_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reports To"
          },
          "runtime_binding_key": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Runtime Binding Key"
          },
          "capability_requirements": {
            "$ref": "#/components/schemas/WorkloadRequirements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role_key",
          "name",
          "responsibility",
          "runtime_binding_key"
        ],
        "title": "AgentRoleSpec",
        "description": "One persistent formal role in an organization definition."
      },
...
      "ErrorEnvelope": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorEnvelope"
      },
...
      "OrganizationDetailResponse": {
        "properties": {
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "current_published_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Published Version Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "current_published_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OrganizationSpec"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "name",
          "description",
          "current_published_version_id",
          "created_at",
          "updated_at",
          "current_published_spec"
        ],
        "title": "OrganizationDetailResponse"
      },
      "OrganizationProposalRequest": {
        "properties": {
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id"
          },
          "source_request": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Request"
          },
          "spec": {
            "$ref": "#/components/schemas/OrganizationSpec"
          }
        },
        "type": "object",
        "required": [
          "spec"
        ],
        "title": "OrganizationProposalRequest"
      },
      "OrganizationSpec": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "roles": {
            "items": {
              "$ref": "#/components/schemas/AgentRoleSpec"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Roles"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "roles"
        ],
        "title": "OrganizationSpecV1",
        "description": "A versioned, portable organization definition for the V1 tree model."
      },
      "OrganizationSummaryResponse": {
        "properties": {
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "current_published_version_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Published Version Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "name",
          "description",
          "current_published_version_id",
          "created_at",
          "updated_at"
        ],
        "title": "OrganizationSummaryResponse"
      },
      "OrganizationVersionResponse": {
        "properties": {
          "spec_version_id": {
            "type": "string",
            "title": "Spec Version Id"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "status": {
            "$ref": "#/components/schemas/OrganizationVersionStatus"
          },
          "spec": {
            "$ref": "#/components/schemas/OrganizationSpec"
          },
          "source_request": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Request"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "spec_version_id",
          "organization_id",
          "version_number",
          "status",
          "spec",
          "source_request",
          "created_at",
          "confirmed_at",
          "published_at"
        ],
        "title": "OrganizationVersionResponse"
      },
      "OrganizationVersionStatus": {
        "type": "string",
        "enum": [
          "proposal",
          "confirmed",
          "published",
          "superseded",
          "archived"
        ],
        "title": "OrganizationVersionStatus"
      },
...
      "WorkloadRequirements": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "1.0",
            "title": "Schema Version",
            "default": "1.0"
          },
          "os_families": {
            "items": {
              "type": "string",
              "enum": [
                "linux",
                "windows",
                "macos"
              ]
            },
            "type": "array",
            "title": "Os Families",
            "default": []
          },
          "requires_gui": {
            "type": "boolean",
            "title": "Requires Gui",
            "default": false
          },
          "requires_gpu": {
            "type": "boolean",
            "title": "Requires Gpu",
            "default": false
          },
          "min_gpu_memory_mb": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Gpu Memory Mb"
          },
          "min_cpu_capacity": {
            "type": "string",
            "enum": [
              "light",
              "standard",
              "heavy"
            ],
            "title": "Min Cpu Capacity",
            "default": "light"
          },
          "min_memory_mb": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Memory Mb"
          },
          "required_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Tools",
            "default": []
          },
          "requires_network": {
            "type": "boolean",
            "title": "Requires Network",
            "default": false
          },
          "required_external_services": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required External Services",
            "default": []
          },
          "required_hardware": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Hardware",
            "default": []
          },
          "required_proprietary_software": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Required Proprietary Software",
            "default": []
          },
          "input_media_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Input Media Types",
            "default": []
          },
          "output_media_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Output Media Types",
            "default": []
          },
          "estimated_duration_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Duration Seconds"
          },
          "estimated_input_size_bytes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Input Size Bytes"
          },
          "resource_intensive": {
            "type": "boolean",
            "title": "Resource Intensive",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "WorkloadRequirementsV1",
        "description": "Capabilities required by one role or one submitted workload."
      }
    }
  }
}